document possible issues with default EHLO argument
By default, msmtp uses "localhost" as an argument to the EHLO command.
Some SMTP servers reject such transactions on the grounds that the
client may be hiding their identity in order to send spam.
Unfortunately, an always-working safe default does not exist and the
user may have to figure out what works for them.

Document that it may be necessary to use the "domain" command under
those circumstances.

Signed-off-by: Clemens Buchacher <drizzd@aon.at>