msmtpq: Create queue directory automatically
If the queue directory is not present, there is no reason to complain as
we can just create it automatically.

This is a quality of life improvement so that people don't have to
remember to create ~/.msmtp.queue on new setups before this script
starts working.

Due to the previous "-d $Q" test, it is only called if the directory
does not exist yet, therefore the "-p" flag is not needed to prevent
mkdir from failing if the directory exists already.

With this patch, this script can finally run with zero prior
configuration.