Don't expect host to never be NULL.
When a socket name is given, the host name may be NULL, which was not yet
accepted everywhere and lead to a null pointer dereference (i.e. crash) in a
strlen() call when logging was active.

In addition to this strlen() call, several printf() calls were fixed, although
these just printed '(null)' instead of dereferencing the pointer.

Reportes as Github issue 79:
https://github.com/marlam/msmtp-mirror/issues/79