msmtpq: inherit MSMTP, Q and LOG from environment
When a user is installing msmtpq simply by copying it to a location on
their PATH, it it perfectly reasonable to ask the user to edit msmtpq
before using it. However, when installing msmtpq via a package manager,
editing the msmtpq executable itself is not a good idea. In general,
files installed by a package manager should only be modified by that
package manager (unless the file is a configuration file).

Additionally, recommending that users to edit a script before it is used
creates friction when the script is updated. Every time the script is
updated, the user has to make those same changes again.

This change allows (and even encourages) the user to set the MSMTP, Q
and LOG variables from outside of the script. This allows users to
configure the location of the msmtp executable, msmtpq’s queue directory
and msmtpq’s log without modifying msmtpq itself.

This change was inspired by a patch that was included in a nixpkgs PR:
<https://github.com/NixOS/nixpkgs/pull/185343/commits/f8e3437ef22573db03e0eb3612373e0e9f2a9257#diff-db16f24325854530247bb84d212795a677d9477a0507fb9aa174f561653c5138R11>