msmtpq: create parent dirs for $Q and $LOG
When I first tried to use msmtpq, I was a little bit confused by the
behavior of the Q and LOG variables. README.msmtpq does say that the
directories for those variables should be created before you run msmtpq,
but I didn’t notice that when I was first skimming through it.

This change makes msmtpq more resilient in those situations. Now, msmtpq
will not only create a directory for $Q and a directory for $LOG but
also create parent directories for those directories if necessary.

This change was inspired by a patch that was included in a nixpkgs PR:
<https://github.com/NixOS/nixpkgs/pull/185343/commits/f8e3437ef22573db03e0eb3612373e0e9f2a9257#diff-db16f24325854530247bb84d212795a677d9477a0507fb9aa174f561653c5138R12>