msmtpq: rename Q to MSMTPQ_Q
8139cf6 (msmtpq: inherit MSMTP, Q and LOG from environment, 2022-08-06)
allowed the Q variable to be set from outside of the msmtpq script.
Before that change, Q was a reasonable name for that variable. Now that
that variable is sharing a namespace with the rest of the system, its
name should reflect the fact that it’s specific to the msmtpq script.

At the time of writing, there’s one release of msmtp (1.8.22) that
contains the changes from 8139cf6 (msmtpq: inherit MSMTP, Q and LOG from
environment, 2022-08-06). For backwards compatibility with that release,
this change makes the Q variable a deprecated alias of MSMTPQ_Q.

I used these scripts [1] to help me test the changes in this commit.

Thanks to Peter Hoeg for suggesting this change [2].

[1]: <https://gist.github.com/Jayman2000/09685f52f3a27450cd65391048b5dd53>
[2]: <https://github.com/marlam/msmtp-mirror/pull/86#discussion_r940283008>