msmtpq: rename LOG to MSMTPQ_LOG
Now that this variable is sharing a namespace with the rest of the
system, its name should reflect the fact that it’s an msmtpq-specific
variable (see the previous commit for details).

I used these scripts [1] to help me test the changes in this commit.

Thanks to Peter Hoeg for suggesting this change [2].

[1]: <https://gist.github.com/Jayman2000/849134cf5d224e864472f779af5050b7>
[2]: <https://github.com/marlam/msmtp-mirror/pull/86#discussion_r940283331>