msmtpq: don’t use log() before it’s defined
To fix this bug, I created a new function called log_later(). I
also considered moving the function definition for log() to before the
variable declarations, moving the variable declarations to underneath
the function declarations and moving the “"$MSMTP" --version” to the
script’s “entry point”. In the end I decided to create log_later() in
order to stay consistent with how the script was already written.

Fixes #93.