macOS portability
The shopt builtin of the bash which ships with macOS does not support
the inherit_errexit option. Hence make sure the statement in line 81
does not fail even if the shopt option is not available.

Similarly, as macOS is a 4.4BSD, and hence its ping's command line
parameters are subtly different from Linux's. The connect_test()
function must make sure to use the appropriate ones.