msmtpq - MSMTPQ_LOG empty string value
The description of MSMTPQ_LOG suggests that having it unset, and
having it set to the empty string were two different cases. The check
in line 162 was not able to distinguish these two cases, however. The
updated check will match when MSMTPQ_LOG is unset only.

In a similar vein, the subsequent check for existence of the log file
directory would try to create a directory even if MSMTPQ_LOG was set
to the empty string, or unset. Added a new if guard to prevent log
directory creation unless a log file will actually be created.