Silence errtrace When Queue Empty
"ls -A" prints a backtrace when the queue directory is empty due to
the errtrace shell option. This commit replaces the call to "ls -A"
with a pure bash function which checks the queue directory for any
message files, and will not trigger a backtrace even when the errtrace
option is set.