Send SMTPUTF8 keyword only when envelope addresses contain non-ASCII domain parts
SMTPUTF8 can cause problems when servers do not implement it, or do not
implement it correctly.  This does not only affect the SMTP server that msmtp
talks to directly, but also servers down the line.

In theory, since msmtp does not inspect the mail before transmission, it should
always send SMTPUTF8 if supported since the message headers or body may require
it.

But in practice this is probably not necessary. For example, postfix always
supports UTF8 in mail header and body, and in envelope address local parts.

Therefore msmtp now only sends the SMTPUTF8 keyword with MAIL FROM when at least
one of the envelope addresses (recipients or sender) contain a non-ascii domain
part.

This should fix (or rather, workaround) the problems reported in Github issue #197.